requirements.txt
asgiref==3.7.2
autopep8==2.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.4.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.2.11
django-allauth==0.63.3
django-storages==1.14.3
exceptiongroup==1.2.0
Faker==12.0.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
mixer==7.2.2
model-bakery==1.18.0
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
